dashboard fixes

+++ model/src/main/kotlin/com/cogoport/ares/model/payment/CollectionTrendResponse.kt
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonAutoDetect
 import io.micronaut.core.annotation.Introspected
 import io.micronaut.core.annotation.ReflectiveAccess
+import java.math.BigDecimal
 
 @Introspected
 @ReflectiveAccess
@@ -13,7 +14,7 @@ data class CollectionTrendResponse(
     @JsonProperty("duration")
     var duration: String?,
     @JsonProperty("receivableAmount")
-    var receivableAmount: Float,
+    var receivableAmount: BigDecimal,
     @JsonProperty("collectableAmount")
-    var collectableAmount: Float
+    var collectableAmount: BigDecimal
 )
