+++ model/src/main/kotlin/com/cogoport/ares/model/payment/DailyOutstandingResponse.kt
@@ -17,5 +17,5 @@ data class DailyOutstandingResponse(
     @JsonProperty("outstandings") var outstandings: BigDecimal?,
     @JsonProperty("totalSales") var totalSales: BigDecimal?,
     @JsonProperty("days") var days: Int,
-    @JsonProperty("value") var value: Float
+    @JsonProperty("value") var value: BigDecimal
 )
